move CertificateInfo back into its own file.  set fingerprints correctly in dtlstest
